requirements.txt
smartapi-python==1.5.5
pyotp
websocket-client
requests
